[#172] REFACT: ArticleDetail, ArticlePrice, MapView 컴포넌트 리팩토링. 불필요한 코드 제거 및 클러스터 관련 유틸리티 함수 개선. 지도 관련 상수 및 스타일 통합.
