+++ src/domain/article/types/article.ts
@@ -1,4 +1,6 @@
 // 매물 응답 타입
+import { REAL_ESTATE_OPTIONS, TRADE_TYPE_OPTIONS } from '../constants/articleConstants';
+
 export type RealEstateType =
     | "아파트"
     | "오피스텔"
@@ -11,22 +13,6 @@ export type RealEstateType =
     | "사무실";
 
 export type TradeType = "매매" | "전세" | "월세";
-
-// 매물 유형 상수
-export const REAL_ESTATE_OPTIONS: RealEstateType[] = [
-    "아파트",
-    "오피스텔",
-    "빌라",
-    "전원주택",
-    "단독/다가구",
-    "상가주택",
-    "한옥주택",
-    "상가",
-    "사무실"
-];
-
-// 거래 유형 상수
-export const TRADE_TYPE_OPTIONS: TradeType[] = ["매매", "전세", "월세"];
 
 export interface ArticleResponse {
     id: number;
